Stvoren novi ekran za instruktor trcanja
Nadodano dohvacanje vjezbe preko korisnikVjezbaDAL
Ispravljeno LIVE dohvacanje Read AtributaKardioVjezbi, te izmjena iz string u int
Ispravak DAO AtributaKardioVjezbi, pocinje se sa 1km zadan, te dobavljanje iz string u int
Nadodan izracun potrosenih kalorija u klasi korisnik
Ispravak modela AtributaKardioVjezbi iz int u float za kalorije potrosene te njegove gettere i settere
Nadograden ViewModel za AtributaKardioVjezbi da dobavlja po Id, da dobavlja vjezbu te broj kalorija
Ispravak bundle dohvacanja u konfiguratoru trcanja
Ispravak MET vrijednosti trcanja

+++ FitApp/core/src/main/java/com/example/core/entities/AtributiKardioVjezbi.java
@@ -26,7 +26,7 @@ public class AtributiKardioVjezbi {
     private float udaljenostPlanirana;
     private float udaljenostOtrcana;
     private int trajanje;
-    private int kalorijaPotroseno;
+    private float kalorijaPotroseno;
 
     public int getId() {
         return id;
@@ -61,11 +61,11 @@ public class AtributiKardioVjezbi {
         this.trajanje = trajanje;
     }
 
-    public int getKalorijaPotroseno() {
+    public float getKalorijaPotroseno() {
         return kalorijaPotroseno;
     }
 
-    public void setKalorijaPotroseno(int kalorijaPotroseno) {
+    public void setKalorijaPotroseno(float kalorijaPotroseno) {
         this.kalorijaPotroseno = kalorijaPotroseno;
     }
 }
